able to build

+++ nRF5_SDK_11.0.0_89a8197/components/libraries/timer/app_timer.h
@@ -279,7 +279,7 @@ uint32_t app_timer_cnt_diff_compute(uint32_t   ticks_to,
  *
  * @return Maximum number of events in queue observed so far.
  */
-uint16_t app_timer_op_queue_utilization_get(void);
+uint8_t app_timer_op_queue_utilization_get(void);
 #endif
 
 #endif // APP_TIMER_H__
